requirements.txt
docutils>=0.18
numpy>=1.18.1
Pillow>=8.4.0
scikit_image>=0.12.3
scipy>=1.4.1
Sphinx>=4.3.0
